Add go-back confirm to viewport

+++ src/component/Viewport.jsx
@@ -1,6 +1,7 @@
 import {Component, h} from 'preact';
 
 import Tube from '@/tube';
+import Prompt from '@/component/Prompt';
 
 import '#/component/Viewport.scss';
 
@@ -12,7 +13,15 @@ export default class Viewport extends Component {
     };
 
     goBack = () => {
-        this.setState({tube: this.state.tube.drop()});
+        if(this.state.tube.front()?.props.confirmLeave) {
+            Prompt.yesNo('leave this page').then(res => {
+                if(res) {
+                    this.setState({tube: this.state.tube.drop()});
+                }
+            });
+        } else {
+            this.setState({tube: this.state.tube.drop()});
+        }
     };
 
     set = view => {
@@ -30,7 +39,7 @@ export default class Viewport extends Component {
                                 <path d="M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20v-2z"/>
                             </svg>
                         </div> 
-                        : ''}
+                    : ''}
                 </div>
                 <div className="view-item">
                     { tube.front() || _.base }
